+++ validator.go
@@ -6,6 +6,8 @@ import (
 	"strings"
 )
 
+type Items map[string]item
+
 type Messages map[string]string
 
 type validatorS struct {
@@ -13,13 +15,19 @@ type validatorS struct {
 	messages Messages
 }
 
-func getErrorMessage(fieldName string, failKey string, messages Messages) string {
+func getErrorMessage(fieldName string, failKey string, options option, messages Messages) string {
 	failKey = strings.ToLower(failKey)
 	if out, ok := messages[failKey]; ok {
-		return out
+		for key, value := range options {
+			out = strings.ReplaceAll(out, "{"+key+"}", value)
+		}
+		return strings.ReplaceAll(out, "{field}", fieldName)
 	} else {
-		if defaultKey, ok := defaultValidatorErrorMessages[failKey]; ok {
-			return fmt.Sprintf(defaultKey, fieldName)
+		if defaultErrorMessage, ok := defaultValidatorErrorMessages[failKey]; ok {
+			for key, value := range options {
+				defaultErrorMessage = strings.ReplaceAll(defaultErrorMessage, "{"+key+"}", value)
+			}
+			return strings.ReplaceAll(defaultErrorMessage, "{field}", fieldName)
 		} else {
 			return fmt.Sprintf("error happened but no error message exists on '%s' item role", failKey)
 		}
@@ -32,22 +40,24 @@ func (o *validatorS) Validate(input interface{}) map[string][]string {
 
 	switch inputValue.Kind() {
 	case reflect.Struct:
-		for fieldName, validator := range o.items {
+		for fieldName, role := range o.items {
 			valueOnKeyInput := inputValue.FieldByName(fieldName)
-			fails := validator.validate(valueOnKeyInput.Interface())
+			fails := role.validate(valueOnKeyInput.Interface())
 			if len(fails) != 0 {
 				for _, failKey := range fails {
-					output[fieldName] = append(output[fieldName], getErrorMessage(fieldName, failKey, o.messages))
+					fieldName = strings.ToLower(fieldName)
+					output[fieldName] = append(output[fieldName], getErrorMessage(fieldName, failKey, role.GetOption(failKey), o.messages))
 				}
 			}
 		}
 	case reflect.Map:
-		for fieldName, validator := range o.items {
+		for fieldName, role := range o.items {
 			valueOnKeyInput := inputValue.MapIndex(reflect.ValueOf(fieldName))
-			fails := validator.validate(valueOnKeyInput.Interface().(string))
+			fails := role.validate(valueOnKeyInput.Interface().(string))
 			if len(fails) != 0 {
 				for _, failKey := range fails {
-					output[fieldName] = append(output[fieldName], getErrorMessage(fieldName, failKey, o.messages))
+					fieldName = strings.ToLower(fieldName)
+					output[fieldName] = append(output[fieldName], getErrorMessage(fieldName, failKey, role.GetOption(failKey), o.messages))
 				}
 			}
 		}