+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(validators): Min and Max function and options parameter added to use in error prints
 * 🎉 feat(utils): determinePrecision function added
 
 .. 1.0.0 (2022-06-22)